feat: Add data source for git_provider. Update ID type of git provider to int64
Issue #2

+++ internal/terrareg/git_provider.go
@@ -6,7 +6,7 @@ import (
 )
 
 type GitProviderModel struct {
-	ID   int    `json:"id" tfsdk:"id"`
+	ID   int64  `json:"id" tfsdk:"id"`
 	Name string `json:"name" tfsdk:"name"`
 }
